Converted to UFR format, rebuilt
Updating font files to the newest version of Playfair Display (1.203), and rebuilt the repository as UFR.

Had to remove a kern pair (gravecomb @guilemot) that was causing the build process to fail. Current version has an *exceptionally* large number of kern pairs, but fontmake handles it.

May be worth reviewing the # of kern pairs for the v2 release.

Fontbakery lists several errors worth noting
- font doesn't need an avar table
- font metrics match previous static instances. Not sure where the "previous version" fontbakery is pulling from.
- I note that the UPM of the old variable version was 1000, but it appears the source is 1240. Should this be reduced?